# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Policy:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.Role:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      rolePolicies:
        items:
          $ref: '#/definitions/models.RolePolicy'
        type: array
      tenantId:
        type: integer
    type: object
  models.RolePolicy:
    properties:
      policy:
        $ref: '#/definitions/models.Policy'
      policyId:
        type: integer
      read:
        type: boolean
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      write:
        type: boolean
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      tenantId:
        type: integer
      username:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is a sample server for using Swagger with Gin.
  title: Swagger Example API
  version: "1.0"
paths:
  /roles:
    get:
      description: Get all roles from the database
      operationId: get-all-roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - Roles
  /roles/delete/{id}:
    delete:
      description: delete a role from the database
      operationId: delete-role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - Roles
  /roles/insert:
    post:
      consumes:
      - application/json
      description: Insert a new role into the database
      operationId: insert-role
      parameters:
      - description: Role to add
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
      security:
      - BearerAuth: []
      summary: Insert a role
      tags:
      - Roles
  /roles/update:
    post:
      consumes:
      - application/json
      description: Update an existing role in the database
      operationId: update-role
      parameters:
      - description: Role to update
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - Roles
  /users:
    get:
      description: get all users with their roles
      operationId: get-all-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Users
  /users/delete/{id}:
    delete:
      description: delete a user from the database
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - Users
  /users/insert:
    post:
      consumes:
      - application/json
      description: insert a new user into the database
      operationId: insert-user
      parameters:
      - description: user to insert
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: Insert a new user
      tags:
      - Users
  /users/update:
    post:
      consumes:
      - application/json
      description: update an existing user in the database
      operationId: update-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: user to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: Update an existing user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"